Warn capath is not supported for websockets
libwebsockets doesn't provide an option to provide a `capath`, ie
a directory that contains multiple certificates.
( https://github.com/warmcat/libwebsockets/issues/3276 )

To avoid confusion, explicitly state that it's not supported for
websockets in the doc for mosquitto.conf, and add a warning if option is
provided while `capath` is not provided.