t/Makefile: Generate test certs if not present in sources
Since generated keys have expiration date,
it means that the tests are not reproductible over time.
Integrator may be tempted to not rely on upstream files
and generate them on the fly at built time.

If need files are present this rule will be skip.

This change was motivated for maintenance of 2.0.11
in Debian 12 (stable).

I noticed that upstream regerated certs in master branch since,
but still they will expire in future.

Origin: https://github.com/eclipse-mosquitto/mosquitto/pull/3234
Relate-to: https://salsa.debian.org/debian-iot-team/mosquitto/-/merge_requests/21
Signed-off-by: Philippe Coval <rzr@users.sf.net>